+++ deadgrep.el
@@ -110,6 +110,9 @@ This is stored as a cons cell of integers (lines-before . lines-after).")
   "We can't guarantee that our process filter will always receive whole lines.
 We save the last line here, in case we need to append more text to it.")
 
+(defvar-local deadgrep--debug-command nil)
+(defvar-local deadgrep--debug-first-output nil)
+
 (defconst deadgrep--position-column-width 5)
 
 (defconst deadgrep--color-code
@@ -210,6 +213,11 @@ We save the last line here, in case we need to append more text to it.")
         (message "Deadgrep finished")))))
 
 (defun deadgrep--process-filter (process output)
+  ;; Searches may see a lot of output, but it's really useful to have
+  ;; a snippet of output when debugging. Store the first output received.
+  (unless deadgrep--debug-first-output
+    (setq deadgrep--debug-first-output output))
+
   ;; If we had an unfinished line from our last call, include that.
   (when deadgrep--remaining-output
     (setq output (concat deadgrep--remaining-output output))
@@ -841,13 +849,15 @@ This will either be a button, a filename, or a search result."
   "Start a ripgrep search."
   (setq deadgrep--spinner (spinner-create 'progress-bar t))
   (spinner-start deadgrep--spinner)
-  (let ((process
-         (start-process-shell-command
-          (format "rg %s" search-term)
-          (current-buffer)
-          (deadgrep--format-command
-           search-term search-type case
-           deadgrep--context))))
+  (let* ((command (deadgrep--format-command
+                   search-term search-type case
+                   deadgrep--context))
+         (process
+          (start-process-shell-command
+           (format "rg %s" search-term)
+           (current-buffer)
+           command)))
+    (setq deadgrep--debug-command command)
     (set-process-filter process #'deadgrep--process-filter)
     (set-process-sentinel process #'deadgrep--process-sentinel)))
 
@@ -936,6 +946,26 @@ for a string, offering the current word as a default."
      search-term
      deadgrep--search-type
      deadgrep--search-case)))
+
+(defun deadgrep-debug ()
+  "Show a buffer with some debug information about the current search."
+  (interactive)
+  (let ((command deadgrep--debug-command)
+        (output deadgrep--debug-first-output)
+        (buf (get-buffer-create "*deadgrep debug*"))
+        (inhibit-read-only t))
+    (pop-to-buffer buf)
+    (erase-buffer)
+    (special-mode)
+    (setq buffer-read-only t)
+
+    (insert
+     "About your environment:\n"
+     (format "Platform: %s\n" system-type)
+     (format "Emacs version: %s\n" emacs-version)
+     (format "Command: %s\n" command)
+     (format "\nInitial output from ripgrep:\n%S" output)
+     (format "\n\nPlease file bugs at https://github.com/Wilfred/deadgrep/issues/new"))))
 
 (provide 'deadgrep)
 ;;; deadgrep.el ends here