+++ CHANGELOG.md
@@ -9,6 +9,10 @@ searches finish really quickly but with zero results.
 Added a `deadgrep-debug` command to help users investigate issues and
 file bugs.
 
+Added a `deadgrep-visit-result-other-window` command (bound to
+<kbd>o</kbd>) for opening results in a separate window. This is handy
+if you want the results buffer to stay visible.
+
 # v0.4
 
 Added the ability to stop searches with `C-c C-k`.